refactor: split GCP snooze mappers into per-component files + specs
Mirror the per-component convention (e.g. aws/prometheus): move the new snooze
mappers out of the grouped monitoring.ts into create_snooze.ts / get_snooze.ts /
expire_snooze.ts, each with its own spec. Shared helper functions stay in
monitoring.ts (now exported); the alerting-policy mappers are left untouched.
monitoring.spec.ts keeps the PromQL alerting-policy test.
